requirements.txt
django
djangorestframework
requests
streamlit
flask
psycopg2-binary
openai